updated how we call createItems so it works in codap v3
Looking at the plugin tutorial (perhaps from 2017 as indicated here) that I worked through in December, https://codap.xyz/pluginTutorialHome.html
it links to example estim8 code and helper code here.
In that code, estim8.js does this, with 1 argument
                pluginHelper.createItems(tCaseValues);      // actually send the data to CODAP
yet pluginHelper.js has this function signature with 2 arguments:
    createItems : function(iValuesArray, iDataContextName) {
But it handles it gracefully if you only pass 1 argument; then iDataContextName is undefined and it checks for that (since I'm a Fortran/C/C++/matlab/Python/R person, I'm surprised it doesn't throw an error and die!)

Since I was building MultiVariateExtras based on code in estim8, I was calling pluginHelper.createItems with just one argument.
But now it looks like the interface has changed: in Tim's more recent code (4 years ago now!) which is what choosy and so MultiVariateExtras uses,
https://github.com/eepsmedia/plugins/blob/master/common/pluginHelper.js
it says
    createItems : function(iValuesArray, iDataContextName, iCallback) {

Confusingly, this even older file from 7 years ago (whose filename claims to be new!)
https://github.com/eepsmedia/plugins/blob/master/common/codap_helper_newAPI.js
had the arguments in a different order,
    createItems : function(iValuesArray, iCallback, iDataContextName) {

Now I'm calling createItems with 3 arguments, each with a name that indicates what it is (to avoid any confusion about argument order).