Added drop-down for legend attribute on plot matrix
It has to include an exception for when we try to make a segmented bar chart, since those already have a legend variable in use (the response variable, which would usually be the y-axis variable). We could make the legend variable a split-at-top attribute for segmented bar charts, but we'll leave that for later. See https://codap.concord.org/how-to/split-a-graph-horizontally-or-vertically-by-a-categorical-attribute/