Added attribute inclusion sliders for Plot Matrix
If we know that an attribute won't be useful on a plot (for example, text-based or categorical with too many categories) it's often nice to not include it when constructing the plot matrix. The table of inclusion sliders also shows the type of each attribute, to help us know what might or might not be useful. We'll have something similar on the Correlation tab soon; there, seeing the attribute type is a reminder to go set them correctly so you get the right kind of correlation/effect size computation (especially when an attribute that is truly numeric doesn't have its type set at all, or set to "none", which then gets treated as categorical)