Removed attributes tab and its source code
The "attributes" tab was left over from the Choosy plugin. It inspired the ability to exclude attributes from the scatterplot matrix, and I re-used its slider/toggles framework and artwork in the Plot Matrix tab (and will use it again in the Correlation tab). But we don't need the "attributes" tab any more, so I removed the html and js associated with it.