Added attribute type and sliders for correlation tab
Sometimes we might want to only compute the correlations/effect sizes for the same attributes that we plotted in the Plot Matrix, so when we plot the correlation matrix it has the same size (number of rows and columns) as the plot matrix. So I added inclusion sliders to the correlation tab, just like the Plot Matrix tab. But, they are independent: if you want, you can set them differently on the two tabs. The table of attributes with sliders also serves as a reminder of the designated type of each attribute, which affects what computation will be done with it. I added a message that users should ensure the types are correct.