Refactored the Pearson correlation computation
computeCorrelationTable was pulling in the data, extracting each pair of numeric columns separately, then sending that to the Pearson correlation computation. That repeated a lot of work. Now it just sends the whole table to the Pearson correlation computation and says which attributes to correlate, avoiding a separate scan through the table and creation of a temporary array.  This will also help when we get to other correlation/effect-size computations.